handle add product to cartdetail and display sum

+++ src/main/java/vn/hoidanit/laptopshop/domain/Cart.java
@@ -20,7 +20,7 @@ public class Cart {
     private long id;
 
     @Min(value = 0)
-    private double sum;
+    private int sum;
 
     @OneToOne()
     @JoinColumn(name = "user_id")
@@ -37,11 +37,11 @@ public class Cart {
         this.id = id;
     }
 
-    public double getSum() {
+    public int getSum() {
         return sum;
     }
 
-    public void setSum(double sum) {
+    public void setSum(int sum) {
         this.sum = sum;
     }
